+++ packages/apps/rwa-demo/src/services/isComplianceOwner.ts
@@ -6,14 +6,14 @@ export interface IIsComplianceOwnerProps {
 }
 
 //@TODO: break this up in seperate owners:
-// only checking the RWA.max-balance-compliance owner.
-// not checking the RWA.supply-limit-compliance
+// only checking the max-balance-compliance owner.
+// not checking the supply-limit-compliance
 // atm I make them the same
 export const isComplianceOwner = async (data: IIsComplianceOwnerProps) => {
   const client = getClient();
 
   const transaction = Pact.builder
-    .execution(`(RWA.max-balance-compliance.is-owner (read-string 'owner))`)
+    .execution(`(max-balance-compliance.is-owner (read-string 'owner))`)
     .setMeta({
       senderAccount: data.owner,
       chainId: getNetwork().chainId,