+++ packages/e2e/e2e-dev-wallet/tests/createAccount.spec.ts
@@ -3,15 +3,18 @@ import { expect } from '@playwright/test';
 
 test('Create account', async ({ initiator, chainweaverApp }) => {
   await test.step('setup', async () => {
-    await chainweaverApp.setup(initiator);
+    await initiator.goto('/');
+    await chainweaverApp.removeSetupProps('initiator');
+    await chainweaverApp.setup(initiator, 'initiator');
+    await initiator.waitForTimeout(1000);
+    await chainweaverApp.selectProfile(initiator, 'Skeletor');
   });
 
   await test.step('create account', async () => {
     await initiator.goto('/');
-    await chainweaverApp.selectNetwork(initiator, 'Development');
-
+    await chainweaverApp.selectNetwork(initiator, 'development');
     const result = await chainweaverApp.createAccount(initiator);
 
-    expect(result).toEqual(true);
+    expect(result).toBeDefined();
   });
 });