+++ packages/libs/kode-ui/src/patterns/CompactTable/CompactTableMobile/CompactTableMobile.tsx
@@ -14,9 +14,10 @@ export const CompactTableMobile: FC<IProps> = ({
   fields,
   isLoading = false,
   variant,
+  ...props
 }) => {
   return data.map((item, idx) => (
-    <section key={idx} className={sectionClass({ variant })}>
+    <section key={idx} className={sectionClass({ variant })} {...props}>
       {fields.map((field, idx) => (
         <div key={`${field.key.toString()}${idx}`} className={rowClass}>
           <span className={headerClass}>{field.label}</span>